% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{bd_as_row}
\alias{bd_as_row}
\title{Convert birth/death event to tibble row}
\usage{
bd_as_row(event_str)
}
\arguments{
\item{event_str}{character vector with format Person (b./d. YYYY)}
}
\value{
tibble with columns Person, Event e.g.

\if{html}{\out{<div class="sourceCode R">}}\preformatted{Person             Event           
Gerardus           Mercator Born On This Day`
Alessandro Volta   Died On This Day
}\if{html}{\out{</div>}}
}
\description{
Convert birth/death event to tibble row
}
