% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wikipedia.R
\name{check_date_input}
\alias{check_date_input}
\title{Check the wiki_date is valid}
\usage{
check_date_input(wiki_date)
}
\arguments{
\item{wiki_date}{string, giving date of desired article.}
}
\value{
multiline string
}
\description{
Check the wiki_date is valid
}
