% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_deaths_table.R
\name{create_births_deaths_table}
\alias{create_births_deaths_table}
\title{create a table of all the people who have been boren/ died on this day}
\usage{
create_births_deaths_table(events_list)
}
\arguments{
\item{events_list}{list of birth/death event strings}
}
\value{
tibble of births and deaths on this day
}
\description{
create a table of all the people who have been boren/ died on this day
}
