% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{create_events_table}
\alias{create_events_table}
\title{Convert list of strings into a tibble of events and details}
\usage{
create_events_table(events_list)
}
\arguments{
\item{events_list}{list of characters}
}
\value{
tbl
}
\description{
Convert list of strings into a tibble of events and details
}
