% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{drop_more_anniveraries}
\alias{drop_more_anniveraries}
\title{Drop all text starting at "More anniversaries"}
\usage{
drop_more_anniveraries(paragraph_str)
}
\arguments{
\item{paragraph_str}{character vector, the wikipedia page content as a single string}
}
\value{
a string with all text between "More anniversaries" and the end excluded
}
\description{
Drop all text starting at "More anniversaries"
}
