% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{drop_substring}
\alias{drop_substring}
\title{Drop all text between start of word and end of the string}
\usage{
drop_substring(paragraph_str, pattern)
}
\arguments{
\item{paragraph_str}{character vector, the wikipedia page content as a single string}

\item{pattern}{character expression to drop}
}
\value{
a string with all text between "More anniversaries" and the end excluded
}
\description{
Drop all text between start of word and end of the string
}
