% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{events_as_row}
\alias{events_as_row}
\title{Convert the yyyy - event pairs into tibble rows}
\usage{
events_as_row(event_str)
}
\arguments{
\item{event_str}{list of character vectors, each containing a yyyy - event pair
e.g. 2012 - A series of blasts occurred at an arms dump ...}
}
\value{
a dataframe with columns: \code{Year} `and Details``
}
\description{
Convert the yyyy - event pairs into tibble rows
}
