% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_daily_facts.R
\name{get_daily_facts}
\alias{get_daily_facts}
\title{Grab daily facts from "https://en.wikipedia.org/wiki/Main_Page" and print out}
\usage{
get_daily_facts(date_str = NULL)
}
\arguments{
\item{date_str}{str, date of article we are looking for.}
}
\description{
Grab daily facts from "https://en.wikipedia.org/wiki/Main_Page" and print out
}
