% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_utils.R
\name{make_pretty_output}
\alias{make_pretty_output}
\title{Factory function to create pretty printing cli functions}
\usage{
make_pretty_output(cli_header, cli_col)
}
\arguments{
\item{cli_header}{cli text object.}

\item{cli_col}{cli colour.}
}
\value{
function to pretty print text to console
}
\description{
Factory function to create pretty printing cli functions
}
\seealso{
headers; \code{\link[cli:cli_h1]{cli::cli_h1()}}, \code{\link[cli:cli_h1]{cli::cli_h2()}}, \code{\link[cli:cli_text]{cli::cli_text()}}; colours
\code{\link[cli:ansi-styles]{cli::col_green()}}, \code{\link[cli:ansi-styles]{cli::col_cyan()}}
}
