% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wikipedia.R
\name{make_request}
\alias{make_request}
\title{Make request to Wikipedia and return page as a string}
\usage{
make_request(wiki_date = NULL)
}
\arguments{
\item{wiki_date}{string, giving date of desired article.}
}
\value{
multiline string
}
\description{
Make request to Wikipedia and return page as a string
}
