% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{remove_footer}
\alias{remove_footer}
\title{Remove all text after "More anniversaries}
\usage{
remove_footer(article_txt)
}
\arguments{
\item{article_txt}{character vector of article text.}
}
\value{
cleaned string
}
\description{
Remove all text after "More anniversaries
}
