% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{remove_substr}
\alias{remove_substr}
\title{Remove substring from article text}
\usage{
remove_substr(article_txt, substr)
}
\arguments{
\item{article_txt}{character vector of article text.}

\item{substr}{pattern to remove.}
}
\value{
cleaned string
}
\description{
Remove substring from article text
}
