% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wikipedia.R
\name{tbl_to_cli_output}
\alias{tbl_to_cli_output}
\title{Convert a tibble into cli list output}
\usage{
tbl_to_cli_output(
  event_tbl,
  header_text,
  title_col = "Year",
  text_col = "Details"
)
}
\arguments{
\item{event_tbl}{tibble, table of events with a date and event string.}

\item{header_text}{character, text for section header.}

\item{title_col}{character, name of col that forms the header.}

\item{text_col}{character, name of event description column.}
}
\description{
Convert a tibble into cli list output
}
