% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_deaths_table.R
\name{text_to_birth_death_list}
\alias{text_to_birth_death_list}
\title{Extract out the births/deaths string as a list - one element per person + date}
\usage{
text_to_birth_death_list(events_list)
}
\arguments{
\item{events_list}{list of character vectors - each a representing a different paragraph.}
}
\value{
list of characters
}
\description{
Extract out the births/deaths string as a list - one element per person + date
}
