% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{text_to_event_list}
\alias{text_to_event_list}
\title{split large string into list of sentences}
\usage{
text_to_event_list(today_list_str)
}
\arguments{
\item{today_list_str}{character vector}
}
\value{
list of characters, each a different sentence
}
\description{
split large string into list of sentences
}
