% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_text.R
\name{to_bd_list}
\alias{to_bd_list}
\title{Extract out the births/deaths string as a list - one element per person + date}
\usage{
to_bd_list(events_list)
}
\arguments{
\item{events_list}{list of characters}
}
\value{
list of characters
}
\description{
Extract out the births/deaths string as a list - one element per person + date
}
