% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wikipedia.R
\name{check_date_input}
\alias{check_date_input}
\title{Check the date_str is valid}
\usage{
check_date_input(date_str)
}
\arguments{
\item{date_str}{string, giving date of desired article.}
}
\value{
multiline string

TODO fix return
}
\description{
Check the date_str is valid
}
