% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wikipedia.R
\name{get_daily_facts}
\alias{get_daily_facts}
\title{Grab daily facts from "https://en.wikipedia.org/wiki/Main_Page" and print out}
\usage{
get_daily_facts(wiki_date = NULL)
}
\arguments{
\item{wiki_date}{str in the format \%d \%B specifying the date of the archived article to retrieve}
}
\description{
Grab daily facts from "https://en.wikipedia.org/wiki/Main_Page" and print out
}
