% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wikipedia.R
\name{read_wiki_html}
\alias{read_wiki_html}
\title{Read in Wikipedia html and convert to string}
\usage{
read_wiki_html(wiki_date = NULL)
}
\arguments{
\item{wiki_date}{str in the format \%d \%B specifying the date of the archived article to retrieve}
}
\value{
character vector with all the wikipedia contents
}
\description{
Read in Wikipedia html and convert to string
}
