% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wikipedia.R
\name{read_wiki_html}
\alias{read_wiki_html}
\title{Read in Wikipedia html and convert to string}
\usage{
read_wiki_html(main_page, archive)
}
\arguments{
\item{main_page}{xmlnode_tree object output from \code{\link[rvest:reexports]{rvest::read_html()}}}
}
\value{
character vector with all the wikipedia contents
}
\description{
Read in Wikipedia html and convert to string
}
